+++ app/src/main/java/com/rikin/hydrohomie/features/hydration/common/domain/HydrationState.kt
@@ -1,9 +1,9 @@
 package com.rikin.hydrohomie.features.hydration.common.domain
 
 data class HydrationState(
-  val drank: Double = 0.0,
-  val goal: Double = 64.0,
-  val drinkAmount: Double = 8.0
+  val drank: Int = 0,
+  val goal: Int = 64,
+  val drinkAmount: Int = 8
 ) {
-  val percent = drank / goal
+  val percent = drank.toFloat() / goal
 }
